+++ models/src/main/java/edu/java/models/dto/response/LinkResponse.java
@@ -1,4 +1,4 @@
 package edu.java.models.dto.response;
 
-public record LinkResponse(int id, String url) {
+public record LinkResponse(long id, String url) {
 }
